refactor(mimo-v2.5): merge mimo_vision into mimo_v2_5 package
The ViT lived in its own one-file mimo_vision/ package used only by MiMo-V2.5.
Fold it into mimo_v2_5/ so the omni model is one directory — matching the
qwen3_omni_moe convention (audio_encoder.py + vision_encoder.py + embedding side
by side). Move vision_encoder.py, drop the mimo_vision package, and repoint the
two lazy imports in embedding.py and the alignment test. Refresh the now-stale
'vision is follow-up / reserved' docstrings.

Also make the V4 fix's jax.sharding import lazy + getattr-guarded: a top-level
'from jax.sharding import NamedSharding' broke the jax-stubbed _prepare_input
payload test (and the test builds the runner via __new__, so self.mesh is unset).

All 12 mimo_v25 test modules (84 tests) green.
