+++ python/sgl_jax/srt/multimodal/models/mimo_v2_5/__init__.py
@@ -1,7 +1,9 @@
-"""MiMo-V2.5 omni model package (audio + text landed; vision is follow-up).
+"""MiMo-V2.5 omni model package (text + audio + image + video understanding).
 
 Exports use lazy attribute loading (PEP 562) so light submodules such as
 ``configuration`` can be imported without pulling the jax/flax model stack.
+The audio tower (``audio_encoder``) and ViT (``vision_encoder``) live alongside
+``embedding`` here; ``embedding`` imports both lazily.
 """
 
 from __future__ import annotations
